refactor: Participant 클래스 상속 적용, 패키지 정리
