+++ src/main/kotlin/blackjack/domain/participant/Hand.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.player
+package blackjack.domain.participant
 
 import blackjack.domain.card.Card
