+++ src/main/kotlin/blackjack/domain/participant/Participant.kt
@@ -1,18 +1,16 @@
-package blackjack.domain.player
+package blackjack.domain.participant
 
 import blackjack.domain.card.Card
 
-data class Player(
-    val name: String
-) {
+open class Participant {
     val hand: Hand = Hand()
-    var status: PlayerStatus = PlayerStatus.HIT
+    var status: ParticipantStatus = ParticipantStatus.HIT
         private set
 
     fun addCards(vararg cards: Card) {
         cards.forEach { hand.add(it) }
         if (hand.isBust()) {
-            changeStatus(PlayerStatus.BUST)
+            changeStatus(ParticipantStatus.BUST)
         }
     }
 
@@ -20,7 +18,7 @@ data class Player(
         return hand.score()
     }
 
-    fun changeStatus(status: PlayerStatus) {
+    fun changeStatus(status: ParticipantStatus) {
         this.status = status
     }
 }
