+++ src/main/kotlin/blackjack/domain/participant/ParticipantStatus.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.participant
+
+enum class ParticipantStatus {
+    STAND, HIT, BUST
+}
