+++ src/main/kotlin/blackjack/domain/participant/Player.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.participant
+
+data class Player(
+    val name: String
+) : Participant()
