--- src/main/kotlin/blackjack/domain/player/PlayerStatus.kt
@@ -1,5 +1,0 @@
-package blackjack.domain.player
-
-enum class PlayerStatus {
-    STAND, HIT, BUST
-}
