+++ src/test/kotlin/blackjack/domain/game/DealerTest.kt
@@ -1,6 +1,7 @@
 package blackjack.domain.game
 
 import blackjack.domain.card.Card
+import blackjack.domain.participant.Dealer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
